… vendor/github.com/golang/groupcache/LICENSE
… vendor/github.com/golang/groupcache/lru/lru.go
… vendor/modules.txt